Move AppearanceMessageWindow to a separate file
