+++ foo_ui_columns/system_appearance_manager.h
@@ -1,0 +1,7 @@
+#pragma once
+
+namespace cui::system_appearance_manager {
+
+void initialise();
+
+}
